SmartTradeAI/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
SQLAlchemy==1.4.25
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
redis==3.5.3
celery==5.2.0
flower==0.9.7
alembic==1.7.4
python-dotenv==0.19.1
yfinance==0.1.63
nsepy==0.8.5
beautifulsoup4==4.10.0
newsapi-python==0.2.6
scikit-learn==1.0.1
prophet==1.0.1
torch==1.10.0
python-telegram-bot==13.7
openai==0.28.1
langgraph==0.0.20
autogen==0.2.9
lightagent==0.1.2
